Rename elevator script to speed controlled sound, and add caching to ComponentFrom* scripts
