+++ src/scripts/scripts/AudioScripts.cc
@@ -53,4 +53,42 @@ namespace sp::scripts {
     };
     StructMetadata MetadataSoundOcclusion(typeid(SoundOcclusion), "SoundOcclusion", "");
     InternalScript<SoundOcclusion> soundOcclusion("sound_occlusion", MetadataSoundOcclusion);
+
+    struct SpeedControlledSound {
+        bool init = false;
+        Transform lastTransform;
+        bool playing = false;
+        int frames = 0;
+        float avgSpeed = 0.0f;
+
+        void OnTick(ScriptState &state, Lock<WriteAll> lock, Entity ent, chrono_clock::duration interval) {
+            if (!ent.Has<TransformSnapshot, Audio>(lock)) return;
+            auto &transform = ent.Get<TransformSnapshot>(lock).globalPose;
+            auto &audio = ent.Get<Audio>(lock);
+
+            if (!init) {
+                lastTransform = transform;
+                init = true;
+            }
+
+            float delta = transform.GetPosition().y - lastTransform.GetPosition().y;
+            bool shouldPlay = std::abs(delta) > 1e-8;
+            if (shouldPlay || frames++ > 69) {
+                if (shouldPlay != playing) {
+                    if (shouldPlay) {
+                        EventBindings::SendEvent(lock, ent, Event{"/sound/play", ent, 0});
+                    } else {
+                        EventBindings::SendEvent(lock, ent, Event{"/sound/stop", ent, 0});
+                    }
+                    playing = shouldPlay;
+                }
+                lastTransform = transform;
+                frames = 0;
+                avgSpeed = 0.9 * avgSpeed + 0.1 * delta;
+                audio.sounds[0].volume = std::min(1.0f, std::abs(avgSpeed) / 0.5f);
+            }
+        }
+    };
+    StructMetadata MetadataSpeedControlledSound(typeid(SpeedControlledSound), "SpeedControlledSound", "");
+    InternalScript<SpeedControlledSound> elevator("speed_controlled_sound", MetadataSpeedControlledSound);
 } // namespace sp::scripts
